Modified student.py and added tests
Inside student.py I modified the __init__ method by adding a new
argument to it and also by checking if the id provided is a valid one or
not. I also added a new method to_dict() to convert all the arguments to
a dictionary.

Inside tests/test_student.py I added tests for instatiating a student
class and also testing the to_dict() method.